+++ src/main/java/com/epam/aidial/cfg/dto/ApplicationEximDto.java
@@ -39,16 +39,5 @@ public class ApplicationEximDto {
     private List<RouteResourceDto> routes;
     private String applicationTypeSchemaId;
     private Map<String, Object> applicationProperties;
-    private McpEximDto mcp;
-
-    @Data
-    public static class McpEximDto {
-        private String endpoint;
-        private TransportDto transport = TransportDto.HTTP;
-        private List<String> allowedTools;
-    }
-
-    public enum TransportDto {
-        HTTP
-    }
+    private McpResourceDto mcp;
 }
