+++ src/main/java/com/epam/aidial/cfg/dto/McpResourceDto.java
@@ -1,0 +1,16 @@
+package com.epam.aidial.cfg.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class McpResourceDto {
+    private String endpoint;
+    private TransportDto transport = TransportDto.HTTP;
+    private List<String> allowedTools;
+
+    public enum TransportDto {
+        HTTP
+    }
+}
