+++ src/main/java/com/epam/aidial/cfg/model/ApplicationExim.java
@@ -39,16 +39,5 @@ public class ApplicationExim {
     private FeaturesResource features;
     private String applicationTypeSchemaId;
     private Map<String, Object> applicationProperties;
-    private Mcp mcp;
-
-    @Data
-    public static class Mcp {
-        private String endpoint;
-        private Transport transport = Transport.HTTP;
-        private List<String> allowedTools;
-    }
-
-    public enum Transport {
-        HTTP
-    }
+    private McpResource mcp;
 }
