+++ src/main/java/com/epam/aidial/cfg/model/ApplicationResource.java
@@ -44,16 +44,5 @@ public class ApplicationResource {
     private String applicationTypeSchemaId;
     private Map<String, Object> applicationProperties;
     private ValidityStateResource validityState;
-    private Mcp mcp;
-
-    @Data
-    public static class Mcp {
-        private String endpoint;
-        private Transport transport = Transport.HTTP;
-        private List<String> allowedTools;
-
-        public enum Transport {
-            HTTP
-        }
-    }
+    private McpResource mcp;
 }
