+++ src/test/java/com/epam/aidial/cfg/functional/tests/ApplicationTypeSchemaFunctionalTest.java
@@ -427,6 +427,11 @@ public abstract class ApplicationTypeSchemaFunctionalTest {
     public void shouldSuccessfullyGetCoreApplicationTypeSchema() {
         typeSchemaFacade.create(dto);
 
+        var mcp = new CoreApplicationTypeSchema.ApplicationTypeMcp();
+        mcp.setEndpoint("http://localhost:9876/mcp");
+        mcp.setAllowedTools(List.of("classify_text"));
+        mcp.setConfigDelivery(CoreApplicationTypeSchema.McpConfigDelivery.META);
+
         CoreApplicationTypeSchema expected = new CoreApplicationTypeSchema();
         expected.setSchema(dto.getSchema());
         expected.setId(dto.getId());
@@ -449,10 +454,6 @@ public abstract class ApplicationTypeSchemaFunctionalTest {
         expected.setDefs(dto.getDefs());
         expected.setProperties(dto.getProperties());
         expected.setRequired(dto.getRequired());
-        var mcp = new CoreApplicationTypeSchema.ApplicationTypeMcp();
-        mcp.setEndpoint("http://localhost:9876/mcp");
-        mcp.setAllowedTools(List.of("classify_text"));
-        mcp.setConfigDelivery(CoreApplicationTypeSchema.McpConfigDelivery.META);
         expected.setApplicationTypeMcp(mcp);
         expected.setApplicationTypeAssistantAttachmentsInRequestSupported(false);
         expected.setApplicationTypeSchemaEndpoint("https://test.com/endpoint");
@@ -568,12 +569,12 @@ public abstract class ApplicationTypeSchemaFunctionalTest {
                       "dial:applicationTypeSchemaEndpoint": "https://test.com/endpoint",
                       "dial:applicationTypeBucketCopy": "ENABLED",
                       "dial:applicationTypeMcp": {
-                                  "dial:endpoint": "http://localhost:9876/mcp",
-                                  "dial:transport": "HTTP",
-                                  "dial:allowedTools": ["classify_text"],
-                                  "dial:mcpConfigDelivery": "META",
-                                  "dial:forwardPerRequestKey": true
-                                },
+                        "dial:endpoint": "http://localhost:9876/mcp",
+                        "dial:transport": "HTTP",
+                        "dial:allowedTools": ["classify_text"],
+                        "dial:mcpConfigDelivery": "META",
+                        "dial:forwardPerRequestKey": true
+                      },
                       "$defs": {
                         "ToolEndpointInfo": {
                           "properties": {
