+++ src/test/resources/app_type_schema_with_routes.json
@@ -66,14 +66,14 @@
   },
   "dial:applicationTypePlaybackSupport": true,
   "dial:appendApplicationPropertiesHeader": true,
-  "dial:applicationTypeAssistantAttachmentsInRequestSupported":false,
+  "dial:applicationTypeAssistantAttachmentsInRequestSupported": false,
   "dial:applicationTypeSchemaEndpoint": "https://test.com/endpoint",
   "dial:applicationTypeMcp": {
-          "dial:endpoint": "http://localhost:9876/mcp",
-          "dial:transport": "HTTP",
-          "dial:allowedTools": ["classify_text"],
-          "dial:mcpConfigDelivery": "META",
-          "dial:forwardPerRequestKey": true
-        },
+    "dial:endpoint": "http://localhost:9876/mcp",
+    "dial:transport": "HTTP",
+    "dial:allowedTools": ["classify_text"],
+    "dial:mcpConfigDelivery": "META",
+    "dial:forwardPerRequestKey": true
+  },
   "$defs": {}
 }
