+++ src/test/resources/application-resources/app_get_dto.json
@@ -28,9 +28,9 @@
     "mcp_toolset": []
   },
   "routes": [],
-  "mcp":{
-    "endpoint":"http://localhost:9876/mcp",
-    "transport":"http",
-    "allowedTools":["classify_text"]
-    }
+  "mcp": {
+    "endpoint": "http://localhost:9876/mcp",
+    "transport": "http",
+    "allowedTools": ["classify_text"]
+  }
 }
