+++ src/test/resources/domain/model/applicationTypeSchemas.json
@@ -16,12 +16,12 @@
     "applicationTypeTruncatePromptEndpoint": "https://example.com/truncate",
     "applicationTypeSchemaEndpoint": "https://test.com/endpoint",
     "applicationTypeMcp": {
-            "endpoint": "http://localhost:9876/mcp",
-            "transport": "http",
-            "allowedTools": ["classify_text"],
-            "mcpConfigDelivery": "meta",
-            "forwardPerRequestKey": true
-          },
+      "endpoint": "http://localhost:9876/mcp",
+      "transport": "http",
+      "allowedTools": ["classify_text"],
+      "mcpConfigDelivery": "meta",
+      "forwardPerRequestKey": true
+    },
     "appendApplicationPropertiesHeader": true,
     "$defs": {
       "maxLength": 255
