+++ src/test/resources/import/import_configWithoutAssistants.json
@@ -80,10 +80,10 @@
       ],
       "endpoint": "endpoint",
       "mcp": {
-          "endpoint": "http://localhost:9876/mcp",
-          "transport": "http",
-          "allowedTools": ["classify_text"]
-        },
+        "endpoint": "http://localhost:9876/mcp",
+        "transport": "http",
+        "allowedTools": ["classify_text"]
+      },
       "defaults": {
         "defaults_key": "defaults_value"
       }
@@ -235,12 +235,12 @@
       "dial:appendApplicationPropertiesHeader": true,
       "dial:applicationTypeSchemaEndpoint": "https://test.com/endpoint",
       "dial:applicationTypeMcp": {
-              "dial:endpoint": "http://localhost:9876/mcp",
-              "dial:transport": "HTTP",
-              "dial:allowedTools": ["classify_text"],
-              "dial:mcpConfigDelivery": "META",
-              "dial:forwardPerRequestKey": true
-            },
+        "dial:endpoint": "http://localhost:9876/mcp",
+        "dial:transport": "HTTP",
+        "dial:allowedTools": ["classify_text"],
+        "dial:mcpConfigDelivery": "META",
+        "dial:forwardPerRequestKey": true
+      },
       "dial:applicationTypeAssistantAttachmentsInRequestSupported": true,
       "$defs": {},
       "properties": {},
