+++ src/test/resources/import/import_preview.json
@@ -395,12 +395,12 @@
         "appendApplicationPropertiesHeader": true,
         "applicationTypeSchemaEndpoint": "https://test.com/endpoint",
         "applicationTypeMcp": {
-                "endpoint": "http://localhost:9876/mcp",
-                "transport": "HTTP",
-                "allowedTools": ["classify_text"],
-                "mcpConfigDelivery": "META",
-                "forwardPerRequestKey": true
-              },
+          "endpoint": "http://localhost:9876/mcp",
+          "transport": "HTTP",
+          "allowedTools": ["classify_text"],
+          "mcpConfigDelivery": "META",
+          "forwardPerRequestKey": true
+        },
         "applicationTypeAssistantAttachmentsInRequestSupported": true,
         "applicationTypeRoutes": [
           {
@@ -943,10 +943,10 @@
           "valid": true
         },
         "mcp": {
-            "endpoint": "http://localhost:9876/mcp",
-            "transport": "http",
-            "allowedTools": ["classify_text"]
-          }
+          "endpoint": "http://localhost:9876/mcp",
+          "transport": "http",
+          "allowedTools": ["classify_text"]
+        }
       }
     }
   ],
