+++ src/test/resources/import/import_zip_preview.json
@@ -322,13 +322,13 @@
         "applicationTypeAssistantAttachmentsInRequestSupported": false,
         "applicationTypeSchemaEndpoint": "https://test.com/endpoint",
         "dial:applicationTypeMcp": {
-                "dial:endpoint": "http://localhost:9876/mcp",
-                "dial:transport": "http",
-                "dial:allowedTools": ["classify_text"],
-                "dial:mcpConfigDelivery": "meta",
-                "dial:forwardPerRequestKey": true
-              },
-        "interceptors":[],
+          "dial:endpoint": "http://localhost:9876/mcp",
+          "dial:transport": "http",
+          "dial:allowedTools": ["classify_text"],
+          "dial:mcpConfigDelivery": "meta",
+          "dial:forwardPerRequestKey": true
+        },
+        "interceptors": [],
         "properties": {},
         "required": [],
         "applicationTypeRoutes": [
@@ -770,10 +770,10 @@
           "valid": true
         },
         "mcp": {
-            "endpoint": "http://localhost:9876/mcp",
-            "transport": "http",
-            "allowedTools": ["classify_text"]
-          }
+          "endpoint": "http://localhost:9876/mcp",
+          "transport": "http",
+          "allowedTools": ["classify_text"]
+        }
       }
     }
   ],
