+++ src/test/resources/import_for_export.json
@@ -49,9 +49,9 @@
       ],
       "endpoint": "endpoint",
       "mcp": {
-          "endpoint": "http://localhost:9876/mcp",
-          "allowedTools": ["classify_text"]
-        }
+        "endpoint": "http://localhost:9876/mcp",
+        "allowedTools": ["classify_text"]
+      }
     }
   },
   "roles": {
@@ -163,11 +163,11 @@
       "dial:applicationTypeInterceptors": ["testInterceptor1"],
       "dial:applicationTypeSchemaEndpoint": "https://test.com/endpoint",
       "dial:applicationTypeMcp": {
-              "dial:endpoint": "http://localhost:9876/mcp",
-              "dial:transport": "HTTP",
-              "dial:allowedTools": ["classify_text"],
-              "dial:mcpConfigDelivery": "META",
-              "dial:forwardPerRequestKey": true
+        "dial:endpoint": "http://localhost:9876/mcp",
+        "dial:transport": "HTTP",
+        "dial:allowedTools": ["classify_text"],
+        "dial:mcpConfigDelivery": "META",
+        "dial:forwardPerRequestKey": true
       },
       "$defs": {},
       "properties": {},
