+++ src/test/resources/import_for_export_dependent_routes.json
@@ -101,10 +101,10 @@
         }
       },
       "mcp": {
-          "endpoint": "http://localhost:9876/mcp",
-          "transport": "http",
-          "allowedTools": ["classify_text"]
-        }
+        "endpoint": "http://localhost:9876/mcp",
+        "transport": "http",
+        "allowedTools": ["classify_text"]
+      }
     }
   },
   "assistant": {
@@ -127,12 +127,12 @@
       "dial:applicationTypeIconUrl": "http://ai-dial-test/openai/deployments/gpt-4o/chat/icon",
       "dial:applicationTypeSchemaEndpoint": "https://test.com/endpoint",
       "dial:applicationTypeMcp": {
-              "dial:endpoint": "http://localhost:9876/mcp",
-              "dial:transport": "HTTP",
-              "dial:allowedTools": ["classify_text"],
-              "dial:mcpConfigDelivery": "META",
-              "dial:forwardPerRequestKey": true
-            },
+        "dial:endpoint": "http://localhost:9876/mcp",
+        "dial:transport": "HTTP",
+        "dial:allowedTools": ["classify_text"],
+        "dial:mcpConfigDelivery": "META",
+        "dial:forwardPerRequestKey": true
+      },
       "dial:applicationTypeRoutes": {
         "route1": {
           "dial:response": {
